# specs/1-confucian-rag-agent/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Confucian RAG Agent API
  version: 0.1.0
servers:
  - url: https://api.example.com
paths:
...
  /provenance/{trace_id}:
    get:
      summary: 查看单次生成的溯源与延迟
      parameters:
        - in: path
          name: trace_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provenance'
        '401': { description: 缺失或无效 API Key }
        '404': { description: 记录不存在或已过期（≥30 天） }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Provenance:
      type: object
      properties:
        trace_id: { type: string }
        model: { type: string }
        prompt_template_version: { type: string }
        sources:
          type: array
          items:
            type: object
            properties:
              passage_id: { type: string }
              title: { type: string }
              location: { type: string }
              confidence: { type: number, format: float, description: Source-level confidence (0..1) }
        params: { type: object }
        latency_ms:
          type: object
          properties:
            text: { type: integer }
            tts: { type: integer }
            render: { type: integer }
            end_to_end: { type: integer }
        confidence:
          type: number
          format: float
          description: Overall confidence (0..1)
...